Added test for delete method

+++ src/main/java/com/forms/personform/service/PersonSrv.java
@@ -19,30 +19,30 @@ public class PersonSrv implements IPersonSrv {
     @Override
     public List<Person> getPeople() {
         List<Person> personList = personDao.getAllPerson();
-        if(personList!=null && !personList.isEmpty()){
-            return  personList;
-        }else{
-            throw  new RuntimeException("No Person not found.");
+        if (personList != null && !personList.isEmpty()) {
+            return personList;
+        } else {
+            throw new RuntimeException("No Person was found.");
         }
     }
 
     @Override
     public List<Person> findByKeyword(String kw) {
         List<Person> personByFistName = personDao.findPersonByFirstName(kw);
-        if(personByFistName!=null && !personByFistName.isEmpty()){
-            return  personByFistName;
-        }else{
-            throw  new RuntimeException("No Person not found.");
+        if (personByFistName != null && !personByFistName.isEmpty()) {
+            return personByFistName;
+        } else {
+            throw new RuntimeException("No Person was found.");
         }
 
     }
 
     @Override
     public Person save(Person person) {
-        if (person!=null){
-            return personDao.insertPerson(person.getFirstname(),person.getLastname(),
-                    person.getBirthday(),person.getAddress());
-        }else {
+        if (person != null) {
+            return personDao.insertPerson(person.getFirstname(), person.getLastname(),
+                    person.getBirthday(), person.getAddress());
+        } else {
             throw new RuntimeException("Can't save a null Person.");
         }
 
@@ -50,7 +50,11 @@ public class PersonSrv implements IPersonSrv {
 
     @Override
     public int delete(Long id) {
-        return personDao.deletePerson(id);
+        if (id != null) {
+            return personDao.deletePerson(id);
+        } else {
+            throw new RuntimeException("The given id is null.");
+        }
     }
 
 }
